third_party/WebKit/LayoutTests/inspector-protocol/timeline/timeline-raf-expected.txt

DIV
Recording started
Tracing complete
RequestAnimationFrame has frame: true
RequestAnimationFrame frames match: true
SUCCESS: found all expected events.
